requirements.txt
celery==5.4.0
redis
django-celery-beat==2.7.0
django-celery-results==2.5.1
python-decouple
psycopg2-binary
